Рефакторит utils
