+++ project/src/components/offer-gallery/offer-gallery.tsx
@@ -1,17 +1,18 @@
 import {Offer} from '../../types/offer';
-import {getRandomNumberOfElements} from '../../utils/utils';
 
 type OfferGalleryProps = {
   offer: Offer;
 };
 
+const MAX_IMAGES = 6;
+
 function OfferGallery({offer}: OfferGalleryProps): JSX.Element {
-  const images = getRandomNumberOfElements(offer.images);
+  const images = offer.images.slice(0, MAX_IMAGES);
 
   return (
     <div className="property__gallery-container container" data-testid="offer-gallery">
       <div className="property__gallery">
-        {images && images.map((source) => (
+        {images.map((source) => (
           <div className="property__image-wrapper" key={source}>
             <img
               className="property__image"
