+++ project/src/utils/utils.ts
@@ -1,0 +1,8 @@
+export const formatFirstLetter = (text: string): string =>
+  text.charAt(0).toUpperCase() + text.slice(1);
+
+export const formatDate = (date: string, locales = 'en-US'): string =>
+  new Date(date).toLocaleString(locales, {month: 'long', year: 'numeric'});
+
+export const calculateRatingWidth = (currentRating: number, maxRating = 5): string =>
+  `${Math.round(currentRating) / maxRating * 100}%`;
